requirements.txt
aniso8601==7.0.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
firebase-admin==4.3.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
grpcio==1.29.0
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
promise==2.3
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rsa==4.0
Rx==1.6.1
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
